Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/workers/Worker_ErrorEvent_message-expected.txt

CONSOLE MESSAGE: line 3: Error: Error Message

PASS  AbstractWorker ErrorEvent.message  
